--- llama-index-integrations/vector_stores/llama-index-vector-store-firestore/llama_index/vector_store/firestore/__init__.py
@@ -1,3 +1,0 @@
-from llama_index.vector_store.firestore.base import FirestoreVectorStore
-
-__all__ = ["FirestoreVectorStore"]
