+++ llama-index-integrations/vector_stores/llama-index-vector-stores-firestore/.gitignore
