+++ llama-index-integrations/vector_stores/llama-index-vector-stores-firestore/llama_index/vector_stores/firestore/__init__.py
@@ -1,0 +1,3 @@
+from llama_index.vector_stores.firestore.base import FirestoreVectorStore
+
+__all__ = ["FirestoreVectorStore"]
